Move common ChefSpec code to aws-parallelcluster-common/spec/spec_helper.rb
